+++ conf.example/envs/metadata_esdb_ipservice.env
@@ -1,5 +1,5 @@
 # Pipeline Configuration
-PIPELINE_YAML=pipelines/metadata_ipservice.yml
+PIPELINE_YAML=pipelines/metadata_esdb_ipservice.yml
 
 # GraphQL Consumer Settings
 GRAPHQL_CONSUMER_ENV_PREFIX=IP_SERVICE
